patched up some bugs that prevent jake's .pro file from compiling on linux
git-svn-id: file:///var/svn/grif/framework@223 d8f7f5d7-83b2-4a66-a7c0-045f5512b77b

+++ trunk/source/GRIXMLParser.cpp
@@ -35,7 +35,7 @@ list<GRIParam*>* GRIXMLParser::readNewParamList(QString filePath, list<GRIParam*
 
                 //debug-------------------------------------------------------
 //                this->pauseProgram("reading tag 'PARAMETER'");
-                  QVariant size = paramList->size();
+                  QVariant size = (unsigned int) paramList->size();
                    paramList->push_back(this->readParameter(xml, size.toString()));
 //                    head->addChild(this->readParameter(xml, head));
             }
